add feature parity with usbkill
now behaves exactly like usbkill, but works in windows, linux and OSX.

Tested on linux and Windows 7 so far (in a VM).

Will shutdown immediately, though it of course requires sudo access on
systems that require it (TODO: Detect this? At the very least make
mention of it in the docs.