use cmd.SysProcAttr in run enumerateDevices()
This change is intended to be used with the app running in "GUI mode",
which currently requires special build flags. It runs the app itself
as well as enumerateDevices()in a "hidden window". This means that,
unlike the Linux/OSX versions of the app, there is no commandline
presence, so the app has to be killed from the task manager. This change
is intended to address issue #5, though there will likely be further
tweaks, so I'm keeping the issue open.